✨ Now corrected the level badge and added ideas
