+++ src/common/playlists/PlayHeaderInfo.jsx
@@ -1,45 +1,9 @@
 import { useState, useEffect } from "react";
-import { RiMedal2Fill } from "react-icons/ri";
-import { IoMdTrophy, IoMdArrowBack } from "react-icons/io";
-import { IoRibbon } from "react-icons/io5";
+import { IoMdArrowBack } from "react-icons/io";
 import { Link } from 'react-router-dom';
 
 import useGitHub from 'common/hooks/useGitHub';
-
-const LevelBadge = ({ level }) => {
-  const [playLevel, setPlayLevel] = useState();
-
-  useEffect(() => {
-    switch (level) {
-      case "Beginner":
-        setPlayLevel(
-          <span className="play-level play-level-1">
-            <IoRibbon size="16px" /> {level}
-          </span>);
-        break;
-      case "Intermediate":
-        setPlayLevel(
-          <span className="play-level play-level-2">
-            <RiMedal2Fill size="16px" /> {level}
-          </span>);
-        break;
-      case "Advanced":
-        setPlayLevel(
-          <span className="play-level play-level-3">
-            <IoMdTrophy size="16px" /> {level}
-          </span>);
-        break;
-      default:
-        setPlayLevel(
-          <span className="play-level play-level-1">
-            <IoRibbon size="16px" /> {level}
-          </span>);
-    }
-  }, [level]);
-  return (
-    <>{playLevel}</>
-  );
-};
+import LevelBadge from 'common/components/LevelBadge';
 
 const Author = ({github}) => {
   const { data, error, isLoading } = useGitHub(github);
